backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
anthropic==0.18.1
httpx==0.27.0
